+++ CHANGELOG.md
@@ -6,6 +6,19 @@ This project follows [Semantic Versioning](http://semver.org/).
 ### Added
 - Added support for adding general extensions.
 
+## [10.3.12] 2025-12-06
+### Fixed
+- Fixed issue with variable deletion not updating the list.
+- Fix invalid source item in enableWhen condition.
+
+## [10.3.11] 2025-12-05
+### Fixed
+- Upgrade to angular 19.2.x and fix npm audit issues.
+
+## [10.3.10] 2025-11-18
+### Added
+- Added support for observation extract extension for valueCode.
+
 ## [10.3.9] 2025-10-13
 ### Fixed
 - Ensure the 'Type initial value' option remains visible when toggling 'Create answer list' or switching items.